+++ src/includes/classes/Actions.php
@@ -66,6 +66,7 @@ class Actions extends AbsBase
      * @since 150422 Rewrite.
      *
      * @param mixed Input action argument(s).
+     * @param mixed $args
      */
     protected function wipeCache($args)
     {
@@ -91,6 +92,7 @@ class Actions extends AbsBase
      * @since 150422 Rewrite.
      *
      * @param mixed Input action argument(s).
+     * @param mixed $args
      */
     protected function clearCache($args)
     {
@@ -116,6 +118,7 @@ class Actions extends AbsBase
      * @since 150422 Rewrite.
      *
      * @param mixed Input action argument(s).
+     * @param mixed $args
      */
     protected function ajaxWipeCache($args)
     {
@@ -140,6 +143,7 @@ class Actions extends AbsBase
      * @since 150422 Rewrite.
      *
      * @param mixed Input action argument(s).
+     * @param mixed $args
      */
     protected function ajaxClearCache($args)
     {
@@ -185,11 +189,10 @@ class Actions extends AbsBase
      * @since 150422 Rewrite.
      *
      * @param mixed Input action argument(s).
+     * @param mixed $args
      */
     protected function saveOptions($args)
     {
-        global $is_apache, $is_nginx;
-
         if (!current_user_can($this->plugin->cap)) {
             return; // Nothing to do.
         } elseif (empty($_REQUEST['_wpnonce']) || !wp_verify_nonce($_REQUEST['_wpnonce'])) {
@@ -222,10 +225,10 @@ class Actions extends AbsBase
             if (!($add_wp_cache_to_wp_config = $this->plugin->addWpCacheToWpConfig())) {
                 $query_args[GLOBAL_NS.'_wp_config_wp_cache_add_failure'] = '1';
             }
-            if ($is_apache && !($add_wp_htaccess = $this->plugin->addWpHtaccess())) {
+            if ($this->plugin->isApache() && !($add_wp_htaccess = $this->plugin->addWpHtaccess())) {
                 $query_args[GLOBAL_NS.'_wp_htaccess_add_failure'] = '1';
             }
-            if ($is_nginx && $this->plugin->applyWpFilters(GLOBAL_NS.'_wp_htaccess_nginx_notice', true)
+            if ($this->plugin->isNginx() && $this->plugin->applyWpFilters(GLOBAL_NS.'_wp_htaccess_nginx_notice', true)
                     && (!isset($_SERVER['WP_NGINX_CONFIG']) || $_SERVER['WP_NGINX_CONFIG'] !== 'done')) {
                 $query_args[GLOBAL_NS.'_wp_htaccess_nginx_notice'] = '1';
             }
@@ -246,7 +249,7 @@ class Actions extends AbsBase
             if (!($remove_wp_cache_from_wp_config = $this->plugin->removeWpCacheFromWpConfig())) {
                 $query_args[GLOBAL_NS.'_wp_config_wp_cache_remove_failure'] = '1';
             }
-            if ($is_apache && !($remove_wp_htaccess = $this->plugin->removeWpHtaccess())) {
+            if ($this->plugin->isApache() && !($remove_wp_htaccess = $this->plugin->removeWpHtaccess())) {
                 $query_args[GLOBAL_NS.'_wp_htaccess_remove_failure'] = '1';
             }
             if (!($remove_advanced_cache = $this->plugin->removeAdvancedCache())) {
@@ -269,11 +272,10 @@ class Actions extends AbsBase
      * @since 150422 Rewrite.
      *
      * @param mixed Input action argument(s).
+     * @param mixed $args
      */
     protected function restoreDefaultOptions($args)
     {
-        global $is_apache, $is_nginx;
-
         if (!current_user_can($this->plugin->cap)) {
             return; // Nothing to do.
         } elseif (is_multisite() && !current_user_can($this->plugin->network_cap)) {
@@ -292,10 +294,10 @@ class Actions extends AbsBase
             if (!($add_wp_cache_to_wp_config = $this->plugin->addWpCacheToWpConfig())) {
                 $query_args[GLOBAL_NS.'_wp_config_wp_cache_add_failure'] = '1';
             }
-            if ($is_apache && !($add_wp_htaccess = $this->plugin->addWpHtaccess())) {
+            if ($this->plugin->isApache() && !($add_wp_htaccess = $this->plugin->addWpHtaccess())) {
                 $query_args[GLOBAL_NS.'_wp_htaccess_add_failure'] = '1';
             }
-            if ($is_nginx && $this->plugin->applyWpFilters(GLOBAL_NS.'_wp_htaccess_nginx_notice', true)
+            if ($this->plugin->isNginx() && $this->plugin->applyWpFilters(GLOBAL_NS.'_wp_htaccess_nginx_notice', true)
                     && (!isset($_SERVER['WP_NGINX_CONFIG']) || $_SERVER['WP_NGINX_CONFIG'] !== 'done')) {
                 $query_args[GLOBAL_NS.'_wp_htaccess_nginx_notice'] = '1';
             }
@@ -316,7 +318,7 @@ class Actions extends AbsBase
             if (!($remove_wp_cache_from_wp_config = $this->plugin->removeWpCacheFromWpConfig())) {
                 $query_args[GLOBAL_NS.'_wp_config_wp_cache_remove_failure'] = '1';
             }
-            if ($is_apache && !($remove_wp_htaccess = $this->plugin->removeWpHtaccess())) {
+            if ($this->plugin->isApache() && !($remove_wp_htaccess = $this->plugin->removeWpHtaccess())) {
                 $query_args[GLOBAL_NS.'_wp_htaccess_remove_failure'] = '1';
             }
             if (!($remove_advanced_cache = $this->plugin->removeAdvancedCache())) {
@@ -341,6 +343,7 @@ class Actions extends AbsBase
      * @since 150422 Rewrite.
      *
      * @param mixed Input action argument(s).
+     * @param mixed $args
      */
     protected function dismissNotice($args)
     {