+++ src/includes/plugin.php
@@ -9,7 +9,7 @@ namespace WebSharks\CometCache;
 use WebSharks\CometCache\Classes;
 
 if (!defined('WPINC')) {
-    exit('Do NOT access this file directly: '.basename(__FILE__));
+    exit('Do NOT access this file directly.');
 }
 require_once __DIR__.'/stub.php';