+++ src/includes/stub.php
@@ -8,12 +8,12 @@
 namespace WebSharks\CometCache;
 
 if (!defined('WPINC')) {
-    exit('Do NOT access this file directly: '.basename(__FILE__));
+    exit('Do NOT access this file directly.');
 }
 require_once dirname(__DIR__).'/vendor/autoload.php';
 require_once __DIR__.'/functions/i18n-utils.php';
 
-${__FILE__}['version'] = '161227'; //version//
+${__FILE__}['version'] = '170201-RC'; //version//
 ${__FILE__}['plugin']  = dirname(dirname(__DIR__));
 ${__FILE__}['plugin'] .= '/'.basename(${__FILE__}['plugin']).'.php';
 ${__FILE__}['ns_path'] = str_replace('\\', '/', __NAMESPACE__); // To dir/path.