+++ uninstall.php
@@ -1,6 +1,6 @@
 <?php
 if (!defined('WPINC')) {
-    exit('Do NOT access this file directly: '.basename(__FILE__));
+    exit('Do NOT access this file directly.');
 }
 $GLOBALS['wp_php_rv'] = '5.4'; //php-required-version//
 if (require(__DIR__.'/src/vendor/websharks/wp-php-rv/src/includes/check.php')) {